[components/masking_functions] PS-9148 postfix: Add caching of dictionary table for component_masking_functions (gen_range fix)
https://perconadev.atlassian.net/browse/PS-9148

Fixed problem with 'gen_range()' UDF not being able to properly handle
negative numbers.

This function internally used to call 'masking_functions::random_number()'
function that used to always accept 'std::size_t' arguments. In case when
negative 'long long' arguments (extracter from UDF context) were passed to
this function, they were implicitly converted to 'std::size_t' resulting
in huge numbers and causing an assertion in 'std::uniform_int_distribution'
constructor that expected 'min' <= 'max'.

Fixed by making the 'masking_functions::random_number()' a function
template.

+++ components/masking_functions/src/masking_functions/dictionary.cpp
@@ -34,7 +34,8 @@ const charset_string &dictionary::get_random() const noexcept {
     return shared_empty;
   }
 
-  const auto random_index{random_number(0, terms_.size() - 1U)};
+  const auto random_index{
+      random_number(static_cast<std::size_t>(0U), terms_.size() - 1U)};
   return *std::next(std::begin(terms_),
                     static_cast<std::ptrdiff_t>(random_index));
 }
